Do not register SmartyStrategy automatically
laminas-mvc registers view strategy in the order they are loaded by
modules. To be able to control the priority of strategies when multiple
strategies are used, the SmartyStrategy is no longer registered
automatically when the module is loaded.

This requires registering the strategy manually in the application's
configuration, but also allows users to register the strategy with a
custom priority.